+++ main/src/main/java/com/coggiri/main/domain/group/repository/GroupRepository.java
@@ -1,11 +1,8 @@
-package com.coggiri.main.mvc.repository;
+package com.coggiri.main.domain.group.repository;
 
 import com.coggiri.main.mvc.domain.dto.GroupInfoDTO;
 import com.coggiri.main.mvc.domain.entity.Group;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.stereotype.Controller;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.util.List;
 import java.util.Map;
