+++ main/src/main/java/com/coggiri/main/domain/user/service/UserService.java
@@ -1,4 +1,4 @@
-package com.coggiri.main.mvc.service;
+package com.coggiri.main.domain.user.service;
 
 import com.coggiri.main.customEnums.Role;
 import com.coggiri.main.mvc.domain.dto.UserDTO;
@@ -6,7 +6,7 @@ import com.coggiri.main.mvc.domain.dto.UserLoginDTO;
 import com.coggiri.main.mvc.domain.entity.JwtToken;
 import com.coggiri.main.mvc.domain.entity.User;
 import com.coggiri.main.mvc.domain.entity.UserGroupRole;
-import com.coggiri.main.mvc.repository.UserRepository;
+import com.coggiri.main.domain.user.repository.UserRepository;
 import com.coggiri.main.jwtUtils.JwtTokenProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
