+++ main/src/main/resources/mybatis/sql/group/Group.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.coggiri.main.mvc.repository.GroupRepository">
+<mapper namespace="com.coggiri.main.domain.group.repository.GroupRepository">
     <insert id="createGroup" parameterType="com.coggiri.main.mvc.domain.entity.Group" useGeneratedKeys="true" keyProperty="groupId" keyColumn="group_id">
         INSERT INTO groups(group_category,group_name,group_intro,group_rule,group_condition,group_number,thumbnail_path)
         Values(
