+++ main/src/main/resources/mybatis/sql/tag/Tag.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.coggiri.main.mvc.repository.TagRepository">
+<mapper namespace="com.coggiri.main.domain.tag.repository.TagRepository">
 
     <insert id="addGroupTag" parameterType="com.coggiri.main.mvc.domain.entity.Tag" useGeneratedKeys="true" keyColumn="tag_id" keyProperty="tags.tagId">
         INSERT INTO group_tag(group_id,tag_name)
