+++ main/src/main/resources/mybatis/sql/task/Task.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.coggiri.main.mvc.repository.TaskRepository">
+<mapper namespace="com.coggiri.main.domain.task.repository.TaskRepository">
     <insert id="createTask" parameterType="com.coggiri.main.mvc.domain.entity.Task" useGeneratedKeys="true" keyColumn="task_id" keyProperty="taskId">
         INSERT INTO task(group_option,achieve_condition,task_name,task_type,register_date,dueDate)
         Values(
