+++ main/src/main/resources/mybatis/sql/user/User.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.coggiri.main.mvc.repository.UserRepository">
+<mapper namespace="com.coggiri.main.domain.user.repository.UserRepository">
     <insert id="register" parameterType="com.coggiri.main.mvc.domain.entity.User" useGeneratedKeys="true" keyProperty="id" keyColumn="id">
         INSERT INTO user(user_id,password,user_name,email)
         Values(
